Move friends to common page
